+++ packages/gatsby-plugin-fastify/src/gatsby-node.ts
@@ -1,6 +1,5 @@
 import { existsSync, mkdir, writeJSON } from "fs-extra";
 
-import type { GatsbyServerFeatureOptions } from "./plugins/gatsby";
 import type { GatsbyNodeServerConfig } from "./utils/config";
 import type { GatsbyNode } from "gatsby";
 
@@ -11,34 +10,34 @@ import { getClientSideRoutes } from "./gatsby/clientSideRoutes";
 
 export const onPostBuild: GatsbyNode["onPostBuild"] = async (
   { store, pathPrefix, reporter },
-  pluginOptions: GatsbyServerFeatureOptions,
+  pluginOptions: GatsbyNodeServerConfig,
 ) => {
   const { redirects } = store.getState();
 
-  const pluginData = await makePluginData(store, pathPrefix);
-
-  const functions = await getFunctionManifest(pluginData);
-  const clientSideRoutes = await getClientSideRoutes(pluginData);
-
-  // @ts-ignore
-  delete pluginOptions.plugins;
-
-  const config: GatsbyNodeServerConfig = {
-    ...pluginOptions,
-    clientSideRoutes,
-    redirects,
-    prefix: pathPrefix,
-    functions,
-  };
-
-  if (!existsSync(PATH_TO_CACHE)) {
-    await mkdir(PATH_TO_CACHE);
-  }
-
   try {
+    const pluginData = await makePluginData(store, pathPrefix);
+
+    const functions = await getFunctionManifest(pluginData);
+    const clientSideRoutes = await getClientSideRoutes(pluginData);
+
+    // @ts-ignore
+    delete pluginOptions.plugins;
+
+    const config: GatsbyNodeServerConfig = {
+      ...pluginOptions,
+      clientSideRoutes,
+      redirects,
+      prefix: pathPrefix,
+      functions,
+    };
+
+    if (!existsSync(PATH_TO_CACHE)) {
+      await mkdir(PATH_TO_CACHE);
+    }
+
     await writeJSON(pluginData.configFolder(CONFIG_FILE_NAME), config, { spaces: 2 });
   } catch (e) {
-    reporter.error("Error writing config file.", e, "gatsby-plugin-fastify");
+    reporter.error("Error building config for Fastify Server", e, "gatsby-plugin-fastify");
   }
 };