+++ packages/gatsby-plugin-fastify/src/plugins/clientPaths.ts
@@ -1,4 +1,7 @@
 import path from "path";
+
+import { PATH_TO_PUBLIC } from "../utils/constants";
+
 import type { FastifyPluginAsync } from "fastify";
 import type { NoUndefinedField } from "../gatsby/clientSideRoutes";
 
@@ -11,7 +14,7 @@ export const handleClientOnlyPaths: FastifyPluginAsync<{
   paths: NoUndefinedField<PathConfig>[];
 }> = async (fastify, { paths }) => {
   for (const p of paths) {
-    console.info("Registering client-only route: ", p.path);
+    fastify.log.info(`Registering client-only route: ${p.path}`);
 
     //TODO: This code only works because I've editted the fastify-static implementation to not encodeURI on file names. https://github.com/fastify/fastify-static/issues/234
     //TODO: Work around for https://github.com/fastify/fastify/issues/3331
@@ -25,7 +28,7 @@ export const handleClientOnlyPaths: FastifyPluginAsync<{
       },
       (_req, reply) => {
         reply.header("x-gatsby-fastify", `served-by: client-only-routes`);
-        reply.sendFile("index.html", path.resolve("./public", p.path.replace("/", "")));
+        reply.sendFile("index.html", path.resolve(PATH_TO_PUBLIC, p.path.replace("/", "")));
       },
     );
   }