+++ packages/gatsby-plugin-fastify/src/plugins/gatsby.ts
@@ -10,22 +10,10 @@ import { getConfig } from "../utils/config";
 import fastifyCompress from "fastify-compress";
 import fastifyAccepts from "fastify-accepts";
 import type { FastifyPluginAsync } from "fastify";
-import type { PluginOptions } from "gatsby";
 
-export interface GatsbyServerFeatureOptions extends PluginOptions {
-  compression: boolean;
-  refreshEndpoint: boolean;
-}
-
-export const serveGatsby: FastifyPluginAsync<GatsbyServerFeatureOptions> = async (fastify) => {
-  const {
-    cli: { verbose },
-    server: serverConfig,
-  } = getConfig();
-
-  if (verbose) {
-    console.info("Starting server with config: ", serverConfig);
-  }
+export const serveGatsby: FastifyPluginAsync = async (fastify) => {
+  //@ts-ignore
+  const { server: serverConfig } = getConfig();
 
   const { clientSideRoutes, redirects, compression, functions } = serverConfig;
 
@@ -34,7 +22,7 @@ export const serveGatsby: FastifyPluginAsync<GatsbyServerFeatureOptions> = async
 
   // Optimizations
   if (compression) {
-    console.info(`Compression enabled.`);
+    fastify.log.info(`Compression enabled`);
     await fastify.register(fastifyCompress, {});
   }