+++ packages/gatsby-plugin-fastify/src/plugins/static.ts
@@ -3,12 +3,13 @@ import fastifyStatic, { FastifyStaticOptions } from "fastify-static";
 import fp from "fastify-plugin";
 import path from "path";
 import { isMatch } from "picomatch";
+import { PATH_TO_PUBLIC } from "../utils/constants";
 
 export const handleStatic: FastifyPluginAsync<Partial<FastifyStaticOptions>> = fp(
   async (fastify, opts) => {
     fastify.register(fastifyStatic, {
-      root: path.resolve("./public"),
-      wildcard: false,
+      root: path.resolve(PATH_TO_PUBLIC),
+      redirect: true,
       setHeaders: (reply, path, _stat) => {
         if (
           isMatch(path, ["**/public/*.@(js|css)", "**/public/static/**"]) &&