+++ packages/gatsby-plugin-fastify/src/serve.ts
@@ -1,28 +1,28 @@
 import { serveGatsby } from "./plugins/gatsby";
 import Fastify from "fastify";
 import { getConfig } from "./utils/config";
-import open from "open";
 
 export async function gatsbyServer() {
   const {
-    cli: { port, host, open: openBrowser, verbose },
+    cli: { port, host, logLevel },
     server: { prefix },
   } = getConfig();
 
-  const fastify = Fastify({ ignoreTrailingSlash: true, logger: verbose });
+  const fastify = Fastify({
+    ignoreTrailingSlash: true,
+    logger: { level: logLevel, prettyPrint: true },
+    disableRequestLogging: ["trace", "debug"].includes(logLevel) ? false : true,
+  });
 
-  console.info("Registered Gatsby @ ", prefix || "/");
-
-  await fastify.register(serveGatsby);
+  fastify.log.info(`Logging Level set @ ${logLevel}`);
+  fastify.log.info(`Mounting Gatsby @ ${prefix || "/"}`);
 
   try {
-    const listeningOn = await fastify.listen(port, host);
+    await fastify.register(serveGatsby, { prefix });
 
-    console.log(`listening @ ${listeningOn}`);
-
-    if (openBrowser) open(listeningOn);
+    await fastify.listen(port, host);
   } catch (err) {
-    console.error("Failed to start Fastify", err);
+    fastify.log.fatal("Failed to start Fastify", err);
     process.exit(1);
   }